requirements.txt
aiohttp==3.12.15
aiofiles==24.1.0
pydantic==2.11.7
m3u8-To-MP4==0.1.11
ffmpeg-python==0.2.0